# server/docs/swagger.yaml
basePath: /
definitions:
  common_schema.CommonAlbumListResp:
    properties:
      cid:
        description: 所属类目
        type: string
      createTime:
        description: 创建时间
        type: string
      ext:
        description: 文件扩展
        type: string
      id:
        description: 主键
        type: string
      name:
        description: 文件名称
        type: string
      path:
        description: 相对路径
        type: string
      size:
        description: 文件大小
        type: string
      updateTime:
        description: 更新时间
        type: string
      uri:
        description: 文件路径
        type: string
    type: object
  common_schema.CommonCateListResp:
    properties:
      createTime:
        description: 创建时间
        type: string
      id:
        description: 主键
        type: string
      name:
        description: 分类名称
        type: string
      pid:
        description: 父级ID
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  common_schema.CommonUploadFileResp:
    properties:
      admin_id:
        description: 管理ID
        type: string
      cid:
        description: 类目ID
        type: string
      ext:
        description: 文件扩展
        type: string
      id:
        description: 主键
        type: string
      name:
        description: 文件名称
        type: string
      path:
        description: 访问地址
        type: string
      size:
        description: 文件大小
        type: integer
      uid:
        description: 用户ID
        type: integer
      url:
        description: 文件路径
        type: string
    type: object
...
  response.PageResp:
    properties:
      count:
        description: 总数
        type: integer
      lists:
        description: 数据列表
      pageNo:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页数量
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: x@adtk.cn
    name: xh
    url: http://x.adtk.cn
  description: x_admin后台管理系统
  license:
    name: MIT License
    url: https://gitee.com/xiangheng/x_admin/blob/main/LICENSE
  termsOfService: http://x.adtk.cn
paths:
...
  /api/admin/common/album/cateAdd:
    post:
      description: 新增相册类目
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 父类目ID
        in: body
        name: pid
        schema:
          type: string
      - description: 类目名称
        in: body
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 类目新增
      tags:
      - common_album-相册管理
  /api/admin/common/album/cateDel:
    post:
      description: 删除相册类目
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 类目ID
        in: body
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 类目删除
      tags:
      - common_album-相册管理
  /api/admin/common/album/cateList:
    get:
      description: 获取相册类目列表
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 类目名称
        in: query
        name: name
        type: string
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/common_schema.CommonCateListResp'
                  type: array
              type: object
      summary: 类目列表
      tags:
      - common_album-相册管理
  /api/admin/common/album/cateRename:
    post:
      description: 相册类目重命名
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 类目ID
        in: body
        name: id
        required: true
        schema:
          type: string
      - description: 新类目名称
        in: body
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 类目重命名
      tags:
      - common_album-相册管理
  /api/admin/common/album/del:
    post:
      description: 删除相册文件
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 文件ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 相册文件删除
      tags:
      - common_album-相册管理
  /api/admin/common/album/list:
    get:
      description: 获取相册文件列表
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 分类ID
        in: query
        name: cid
        type: string
      - description: 文件名
        in: query
        name: name
        type: string
      - collectionFormat: csv
        description: 文件扩展名
        in: query
        items:
          type: string
        name: ext
        type: array
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResp'
                  - properties:
                      lists:
                        $ref: '#/definitions/common_schema.CommonAlbumListResp'
                    type: object
              type: object
      summary: 相册文件列表
      tags:
      - common_album-相册管理
  /api/admin/common/album/move:
    post:
      description: 相册文件移动到指定分类
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 文件ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      - description: 目标分类ID
        in: body
        name: cid
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 相册文件移动
      tags:
      - common_album-相册管理
  /api/admin/common/album/rename:
    post:
      description: 相册文件重命名
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 文件ID
        in: body
        name: id
        required: true
        schema:
          type: string
      - description: 新文件名
        in: body
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 相册文件重命名
      tags:
      - common_album-相册管理
...
  /api/admin/common/index/config:
    get:
      description: 获取系统公共配置
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 公共配置
      tags:
      - common_index-首页
  /api/admin/common/index/console:
    get:
      description: 获取控制台统计数据
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 控制台数据
      tags:
      - common_index-首页
  /api/admin/common/upload/preUpload:
    post:
      description: 文件预上传检查
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 文件MD5
        in: body
        name: md5
        required: true
        schema:
          type: string
      - description: 文件名
        in: body
        name: fileName
        required: true
        schema:
          type: string
      - description: 文件大小
        in: body
        name: fileSize
        required: true
        schema:
          type: integer
      - description: 分类ID
        in: body
        name: cid
        schema:
          type: string
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 文件预上传
      tags:
      - common_upload-上传
  /api/admin/common/upload/upload:
    post:
      description: 上传文件
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 分类ID
        in: body
        name: cid
        schema:
          type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/common_schema.CommonUploadFileResp'
              type: object
      summary: 上传文件
      tags:
      - common_upload-上传
...
swagger: "2.0"